Refactor NavigatorIOS to <Avigator> abstraction
