+++ src/lib/Scenes/MyCollection/Screens/ArtworkDetail/MyCollectionArtworkDetail.tsx
@@ -11,13 +11,11 @@ import { defaultEnvironment } from "lib/relay/createEnvironment"
 import { ScreenMargin } from "lib/Scenes/MyCollection/Components/ScreenMargin"
 import { AppStore } from "lib/store/AppStore"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
-import { useScreenDimensions } from "lib/utils/useScreenDimensions"
 import { Button, Join, Separator, Spacer } from "palette"
 import React from "react"
 import { ScrollView } from "react-native"
-import NavigatorIOS from "react-native-navigator-ios"
 import { graphql, QueryRenderer } from "react-relay"
-import { useNavigator } from "../../utils/useNavigator"
+import { Navigator } from "../../Components/Navigator"
 import { MyCollectionArtworkInsightsFragmentContainer as ArtworkInsights } from "./Components/ArtworkInsights/MyCollectionArtworkInsights"
 import { MyCollectionArtworkHeaderFragmentContainer as ArtworkHeader } from "./Components/MyCollectionArtworkHeader"
 import { MyCollectionArtworkMetaFragmentContainer as ArtworkMeta } from "./Components/MyCollectionArtworkMeta"
@@ -31,54 +29,42 @@ export interface MyCollectionArtworkDetailProps {
 const MyCollectionArtworkDetail: React.FC<MyCollectionArtworkDetailProps> = ({ artwork, marketPriceInsights }) => {
   const navActions = AppStore.actions.myCollection.navigation
   const artworkActions = AppStore.actions.myCollection.artwork
-  const dimensions = useScreenDimensions()
 
   return (
-    <NavigatorIOS
-      navigationBarHidden={true}
-      style={{ height: dimensions.height }}
-      initialRoute={{
-        title: "",
-        component: ({ navigator }) => {
-          useNavigator(navigator)
+    <Navigator>
+      <ScrollView>
+        <FancyModalHeader
+          leftButtonText=""
+          rightButtonText="Edit"
+          onRightButtonPress={() => {
+            artworkActions.startEditingArtwork(artwork as any) // FIXME: remove `any` type
+          }}
+          hideBottomDivider
+        />
+        <Join separator={<Spacer my={1} />}>
+          <ArtworkHeader artwork={artwork} />
+          <ArtworkMeta artwork={artwork} />
+          <Separator />
+          <ArtworkInsights artwork={artwork} marketPriceInsights={marketPriceInsights} />
+          <Separator />
+          <WhySell />
 
-          return (
-            <ScrollView>
-              <FancyModalHeader
-                leftButtonText=""
-                rightButtonText="Edit"
-                onRightButtonPress={() => {
-                  artworkActions.startEditingArtwork(artwork as any) // FIXME: remove `any` type
-                }}
-                hideBottomDivider
-              />
-              <Join separator={<Spacer my={1} />}>
-                <ArtworkHeader artwork={artwork} />
-                <ArtworkMeta artwork={artwork} />
-                <Separator />
-                <ArtworkInsights artwork={artwork} marketPriceInsights={marketPriceInsights} />
-                <Separator />
-                <WhySell />
+          <ScreenMargin>
+            <Button size="large" block onPress={() => navActions.navigateToConsign()}>
+              Submit this work
+            </Button>
 
-                <ScreenMargin>
-                  <Button size="large" block onPress={() => navActions.navigateToConsign()}>
-                    Submit this work
-                  </Button>
+            <Spacer my={0.5} />
 
-                  <Spacer my={0.5} />
+            <Button size="large" variant="secondaryGray" block>
+              Learn more
+            </Button>
+          </ScreenMargin>
 
-                  <Button size="large" variant="secondaryGray" block>
-                    Learn more
-                  </Button>
-                </ScreenMargin>
-
-                <Spacer my={2} />
-              </Join>
-            </ScrollView>
-          )
-        },
-      }}
-    />
+          <Spacer my={2} />
+        </Join>
+      </ScrollView>
+    </Navigator>
   )
 }
