--- src/lib/Scenes/MyCollection/utils/useNavigator.ts
@@ -1,13 +1,0 @@
-import { AppStore } from "lib/store/AppStore"
-import { useEffect } from "react"
-import NavigatorIOS from "react-native-navigator-ios"
-
-export function useNavigator(newNavigator: NavigatorIOS) {
-  const navActions = AppStore.actions.myCollection.navigation
-
-  useEffect(() => {
-    navActions.setNavigator(newNavigator)
-  }, [])
-
-  return newNavigator
-}
